Move FormattingContext creation.
- Moved its creation from within the `RazorFormattingService` into the `FormattingContext` class itself.
- Needed to move the `FormattingContext` creation into a singular place to allow for `OnAutoInsert` handlers to use it. Decided to avoid using a service because the creation of the context was pretty contained (could be static).
- Did not add specific tests for the `FormattingContext` because we didn't have them before. Instead the contexts were always tested indirectly via more integration like tests.

Part of dotnet/aspnetcore#22716
