+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Formatting/DefaultRazorFormattingService.cs
@@ -3,11 +3,9 @@
 
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Razor.Language;
-using Microsoft.AspNetCore.Razor.Language.Legacy;
 using Microsoft.AspNetCore.Razor.Language.Syntax;
 using Microsoft.AspNetCore.Razor.LanguageServer.Common;
 using Microsoft.CodeAnalysis.Text;
@@ -37,6 +35,11 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Formatting
             ILanguageServer server,
             ILoggerFactory loggerFactory)
         {
+            if (formatOnTypeProviders is null)
+            {
+                throw new ArgumentNullException(nameof(formatOnTypeProviders));
+            }
+
             if (documentMappingService is null)
             {
                 throw new ArgumentNullException(nameof(documentMappingService));
@@ -111,7 +114,7 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Formatting
                 return Task.FromResult(EmptyArray);
             }
 
-            var formattingContext = CreateFormattingContext(uri, codeDocument, new Range(position, position), options);
+            var formattingContext = FormattingContext.Create(uri, codeDocument, new Range(position, position), options);
             for (var i = 0; i < applicableProviders.Count; i++)
             {
                 if (applicableProviders[i].TryFormatOnType(position, formattingContext, out var textEdits))
@@ -148,7 +151,7 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Formatting
                 throw new ArgumentNullException(nameof(options));
             }
 
-            var formattingContext = CreateFormattingContext(uri, codeDocument, range, options);
+            var formattingContext = FormattingContext.Create(uri, codeDocument, range, options);
             var edits = await FormatCodeBlockDirectivesAsync(formattingContext);
             return edits;
         }
@@ -507,97 +510,6 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Formatting
             var precedingWhitespaceLength = precedingLineText.GetTrailingWhitespace().Length;
 
             return TextSpan.FromBounds(start - precedingWhitespaceLength, end);
-        }
-
-        // Internal for testing
-        internal static FormattingContext CreateFormattingContext(Uri uri, RazorCodeDocument codedocument, Range range, FormattingOptions options)
-        {
-            var result = new FormattingContext()
-            {
-                Uri = uri,
-                CodeDocument = codedocument,
-                Range = range,
-                Options = options
-            };
-
-            var source = codedocument.Source;
-            var syntaxTree = codedocument.GetSyntaxTree();
-            var formattingSpans = syntaxTree.GetFormattingSpans();
-
-            var total = 0;
-            var previousIndentationLevel = 0;
-            for (var i = 0; i < source.Lines.Count; i++)
-            {
-                // Get first non-whitespace character position
-                var lineLength = source.Lines.GetLineLength(i);
-                var nonWsChar = 0;
-                for (var j = 0; j < lineLength; j++)
-                {
-                    var ch = source[total + j];
-                    if (!char.IsWhiteSpace(ch) && !ParserHelpers.IsNewLine(ch))
-                    {
-                        nonWsChar = j;
-                        break;
-                    }
-                }
-
-                // position now contains the first non-whitespace character or 0. Get the corresponding FormattingSpan.
-                if (TryGetFormattingSpan(total + nonWsChar, formattingSpans, out var span))
-                {
-                    result.Indentations[i] = new IndentationContext
-                    {
-                        Line = i,
-                        IndentationLevel = span.IndentationLevel,
-                        RelativeIndentationLevel = span.IndentationLevel - previousIndentationLevel,
-                        ExistingIndentation = nonWsChar,
-                        FirstSpan = span,
-                    };
-                    previousIndentationLevel = span.IndentationLevel;
-                }
-                else
-                {
-                    // Couldn't find a corresponding FormattingSpan.
-                    result.Indentations[i] = new IndentationContext
-                    {
-                        Line = i,
-                        IndentationLevel = -1,
-                        RelativeIndentationLevel = previousIndentationLevel,
-                        ExistingIndentation = nonWsChar,
-                    };
-                }
-
-                total += lineLength;
-            }
-
-            return result;
-        }
-
-        private static bool TryGetFormattingSpan(int absoluteIndex, IReadOnlyList<FormattingSpan> formattingspans, out FormattingSpan result)
-        {
-            result = null;
-            for (var i = 0; i < formattingspans.Count; i++)
-            {
-                var formattingspan = formattingspans[i];
-                var span = formattingspan.Span;
-
-                if (span.Start <= absoluteIndex)
-                {
-                    if (span.End >= absoluteIndex)
-                    {
-                        if (span.End == absoluteIndex && span.Length > 0)
-                        {
-                            // We're at an edge.
-                            // Non-marker spans (spans.length == 0) do not own the edges after it
-                            continue;
-                        }
-
-                        result = formattingspan;
-                        return true;
-                    }
-                }
-            }
-
-            return false;
         }
     }
 }
